requirements.txt
pandas>=2.2.3
numpy>=1.26.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
python-dotenv==1.0.1
streamlit>=1.40.0
pydantic>=2.9.0